third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-size-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-size value '1px'
  assert_equals: expected "1px auto" but got "1px"
[FAIL] Property background-size value '1px auto'
  assert_equals: expected "1px auto" but got "1px"
Harness: the test ran to completion.
